@@ -104,24 +107,19 @@
     "style-loader": "^0.18.2",
     "superagent": "^3.6.0",
     "supertest": "^3.0.0",
-    "validator": "^5.4.0",
-    "webpack": "^2.7.0",
+    "swagger-jsdoc": "^1.9.7",
+    "uglifyjs-webpack-plugin": "^1.0.1",
+    "validator": "^5.7.0",
+    "webpack": "^3.8.1",
     "webpack-dev-middleware": "^1.12.0",
     "webpack-dev-server": "^2.8.2",
     "webpack-hot-middleware": "^2.18.2",
+    "webpack-merge": "^4.1.1",
     "whatwg-fetch": "^2.0.3",
-    "worker-loader": "^1.0.0"
+    "worker-loader": "^1.1.0"
   },
   "devDependencies": {
-    "babel-cli": "^6.24.1",
     "babel-jest": "^21.2.0",
-    "babel-polyfill": "^6.23.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-2": "^6.24.1",
-    "babel-watch": "^2.0.7",
-    "chai": "^4.1.1",
-    "chai-http": "^3.0.0",
     "coveralls": "^2.13.1",
     "eslint": "^4.3.0",
     "eslint-config-airbnb": "^15.1.0",
