@@ -1,16 +1,11 @@
-import axios from 'axios';
 import { browserHistory } from 'react-router';
 import jwtDecode from 'jwt-decode';
-
 import { changepasswordError, changepasswordRequest, changepasswordResponse } from '../../actions/changePassword';
-
-
+import { validatePasswordChange } from '../components/validationHelperFunctions/validatePasswordChange';
 /**
- * 
- * 
  * @export
- * @param {any} userData 
- * @param {any} token 
+ * @param {object} userData 
+ * @param {string} token 
  * @returns 
  */
 export default function changePassword(userData, token) {
@@ -18,30 +13,34 @@ export default function changePassword(userData, token) {
   let userId = decodedToken.id;
   return (dispatch) => {
     dispatch(changepasswordRequest(userData));
-
-    return fetch('/api/v2/users/' + userId + '/password', {
-      method: 'PUT',
-      headers: {
-        'Accept': 'application/json, text/plain, */*',
-        'Content-Type': 'application/json',
-        token: token
-      },
-      body: JSON.stringify(userData)
-    })
-      .then(
-      (res) => res.json())
-      .then((response) => {
-        if (response.status >= 400) {
-          throw response.message
+    validatePasswordChange(userData)
+      .then((responseMessage) => {
+        if (responseMessage !== '') {
+          dispatch(changepasswordError(responseMessage));
         }
-        else if (response.status === 200) {
-          dispatch(changepasswordResponse(response.message));
+        else {
+          return fetch('/api/v2/users/' + userId + '/password', {
+            method: 'PUT',
+            headers: {
+              'Accept': 'application/json, text/plain, */*',
+              'Content-Type': 'application/json',
+              token: token
+            },
+            body: JSON.stringify(userData)
+          })
+            .then(
+            (res) => {
+              if (res.status >= 400) {
+                res.json().then((response) => {
+                  dispatch(changepasswordError(response.message))
+                })
+              } else {
+                res.json().then((response) => {
+                  dispatch(changepasswordResponse(response.message));
+                })
+              }
+            })
         }
       })
-      .catch(error => {
-        dispatch(changepasswordError(error))
-      });
-
   }
-
 }
