@@ -1,21 +1,12 @@
 import axios from 'axios';
-import {
-  browserHistory
-} from 'react-router';
+import { browserHistory } from 'react-router';
 import 'whatwg-fetch'
-
-import {
-  getbooksError,
-  getbooksRequest,
-  getbooksReponse
-} from '../../actions/getBooks';
+import { getbooksError, getbooksRequest, getbooksReponse } from '../../actions/getBooks';
 
 /**
- * 
- * 
  * @export
- * @param {any} set 
- * @returns 
+ * @param {boolean} set 
+ * @returns {object} response
  */
 export default function getbooks(set) {
   const token = window.localStorage.jwtToken;
@@ -23,6 +14,10 @@ export default function getbooks(set) {
     dispatch(getbooksRequest());
     return fetch('/api/v2/books', {
       method: 'GET',
+      headers: {
+        'Accept': 'application/json, text/plain, */*',
+        'Content-Type': 'application/json'
+      },
       body: {
         token: token
       },
@@ -31,21 +26,19 @@ export default function getbooks(set) {
       },
     })
       .then(
-      (res) => res.json())
-      .then((response) => {
-        if (response.status >= 400) {
-
-          throw response.message
-        } else if (response.status === 200) {
-          dispatch(getbooksReponse(response.books));
-          if (!set) {
-            browserHistory.push('/books');
-          } else if (set) { }
+      (res) => {
+        if (res.status >= 400) {
+          res.json().then((response) => {
+            dispatch(getbooksError(response.message))
+          })
+        } else {
+          res.json().then((response) => {
+            dispatch(getbooksReponse(response.books));
+            if (!set) {
+              browserHistory.push('/books');
+            } else if (set) { }
+          })
         }
-      })
-      .catch(
-      error => {
-        dispatch(getbooksError(error))
       }
       )
   };
