@@ -32,19 +32,18 @@ export default function returnbook(bookData, token) {
       body: JSON.stringify(bookData)
     })
       .then(
-      (res) => res.json())
-      .then((response) => {
-        if (response.status >= 400) {
-          throw response.message
+      (res) => {
+        if (res.status >= 400) {
+          res.json().then((response) => {
+            dispatch(returnbookError(response.message));
+          })
         }
-        else if (response.status === 200) {
-          dispatch(returnbookResponse(response.message));
+        else {
+          res.json().then((response) => {
+            dispatch(returnbookResponse(response.message));
+          })
         }
       })
-      .catch(error => {
-        dispatch(returnbookError(error));
-      });
-
   }
 
 }
