@@ -1,26 +1,21 @@
-import axios from 'axios';
 import { browserHistory } from 'react-router';
 import jwtDecode from 'jwt-decode';
 
 import { uploadImage } from './uploadImage';
 import { updateuserError, updateuserRequest, updateuserResponse } from '../../actions/updateUser';
-
+import { validateUserDetailsUpdate } from '../components/validationHelperFunctions/validateUserDetailsUpdate';
 
 /**
- * 
- * 
  * @export
- * @param {any} userData 
- * @param {any} token 
+ * @param {object} userData 
+ * @param {string} token 
  * @returns 
  */
 export default function updateUser(userData, token) {
   let decodedToken = jwtDecode(token);
   let userId = decodedToken.id;
-
   return (dispatch) => {
     dispatch(updateuserRequest(userData));
-
     return fetch('/api/v2/users/' + userId, {
       method: 'PUT',
       headers: {
@@ -31,18 +26,17 @@ export default function updateUser(userData, token) {
       body: JSON.stringify(userData)
     })
       .then(
-      (res) => res.json())
-      .then((response) => {
-        if (response.status >= 400) {
-          throw response.message
+      (res) => {
+        if (res.status >= 400) {
+          res.json().then((response) => {
+            dispatch(updateuserError(response.message));
+          })
+        } else {
+          res.json().then((response) => {
+            dispatch(updateuserResponse(response.message));
+          })
         }
-        else if (response.status === 200) {
-          dispatch(updateuserResponse(response.message));
-        }
-      })
-      .catch(error => {
-        dispatch(uupdateuserError(error));
-      });
+      }
+      )
   }
-
 }
