@@ -1,22 +1,11 @@
 import React from 'react';
-import lodash from 'lodash';
+import _ from 'lodash';
 import $ from 'jquery';
 import PropTypes from 'prop-types';
 import ActivityLoader from '../preloader/ActivityLoader';
-
-
-let bookId = '';
-let filteredData = '';
-let bookTitle = '';
-let id = '';
-let category = '';
-let isbn = '';
-let stocknumber = 0;
-let author = '';
-let image = '';
-let summary = '';
-let sortedData = '';
-let display = 'none';
+import { storeBookDetailsTemp } from '../BookDetail/helperFunctions/storeBookDetailsTemp';
+import { getStoredBookDetails } from '../BookDetail/helperFunctions/getStoredBookDetails';
+import { removeStoredBookDetails } from '../BookDetail/helperFunctions/removeStoredBookDetails';
 
 /**
  * @class DetailsForm
@@ -34,8 +23,11 @@ class DetailsForm extends React.Component {
       message: '',
       error: '',
       display: false,
+      displayPreloader: 'none',
+      filteredData: {},
+      bookData: [],
+      bookIndex: 0
     };
-
     this.handleClick = this.handleClick.bind(this);
     this.handleOpen = this.handleOpen.bind(this);
     this.handleClose = this.handleClose.bind(this);
@@ -46,51 +38,37 @@ class DetailsForm extends React.Component {
    * @returns {void}
    */
   componentWillMount() {
-    if (!lodash.isEmpty(localStorage.bookId)) {
+    if (!_.isEmpty(localStorage.bookId)) {
       this.props.checkBookDetails(localStorage.bookId);
     }
-
-    if (!lodash.isEmpty(this.props.book[0])) {
-      if (lodash.isEmpty(localStorage.id)) {
+    if (!_.isEmpty(this.props.book[0])) {
+      if (_.isEmpty(localStorage.id)) {
         const { bookData } = this.props;
-        bookId = this.props.book[0].bookId;
+        const bookId = this.props.book[0].bookId;
         window.localStorage.setItem('bookId', bookId);
-        filteredData = lodash
-          .filter(bookData, bookItem => bookItem.id.toString() === bookId.toString());
-        bookTitle = filteredData[0].bookTitle;
-        id = filteredData[0].id;
-        category = filteredData[0].category;
-        isbn = filteredData[0].isbn;
-        stocknumber = filteredData[0].stocknumber;
-        author = filteredData[0].author;
-        image = filteredData[0].image;
-        summary = filteredData[0].summary;
-        localStorage.setItem('image', image);
-        localStorage.setItem('bookTitle', bookTitle);
-        localStorage.setItem('id', id);
-        localStorage.setItem('category', category);
-        localStorage.setItem('isbn', isbn);
-        localStorage.setItem('stocknumber', stocknumber);
-        localStorage.setItem('author', author);
-        localStorage.setItem('summary', summary);
+        const bookIndex = _.findIndex(bookData, ['id', Number(bookId)]);
+        const filteredData = bookData[bookIndex];
+        this.setState({
+          filteredData,
+          bookIndex,
+          bookData
+        });
       }
     }
-
-
     if (this.props.bookData.length === 0) {
-      bookTitle = localStorage.getItem('bookTitle');
-      id = localStorage.getItem('id');
-      category = localStorage.getItem('category');
-      isbn = localStorage.getItem('isbn');
-      stocknumber = localStorage.getItem('stocknumber');
-      author = localStorage.getItem('author');
-      summary = localStorage.getItem('summary');
-      image = localStorage.getItem('image');
+      getStoredBookDetails(this.state.filteredData);
     }
 
     $(document).ready(() => {
       $('.modal').modal();
     });
+  }
+  /**
+  * @memberof DetailsForm
+  * @returns {void}
+  */
+  componentDidMount() {
+    storeBookDetailsTemp(this.state.filteredData);
   }
 
   /**
@@ -99,17 +77,19 @@ class DetailsForm extends React.Component {
    * @returns {void}
    */
   componentWillReceiveProps(nextProps) {
-    sortedData = nextProps.borrowBookItem[0];
-
-    display = 'none';
+    const sortedData = nextProps.borrowBookItem[0];
+    this.setState({
+      displayPreloader: 'none'
+    });
     if (this.state.display) {
-      if (!lodash.isEmpty(sortedData.error) && this.state.display) {
+      if (!_.isEmpty(sortedData.error) && this.state.display) {
         this.setState({
           error: sortedData.error,
           display: false,
         });
         $('#modal2').modal('open');
-      } else if (!lodash.isEmpty(sortedData.response) && this.state.display) {
+      } else if (!_.isEmpty(sortedData.response) && this.state.display) {
+        this.state.bookData[this.state.bookIndex].stockNumber -= 1;
         this.setState({
           message: sortedData.response,
           display: false,
@@ -124,15 +104,7 @@ class DetailsForm extends React.Component {
    * @returns {void}
    */
   componentWillUnmount() {
-    localStorage.removeItem('bookId');
-    localStorage.removeItem('id');
-    localStorage.removeItem('category');
-    localStorage.removeItem('isbn');
-    localStorage.removeItem('stocknumber');
-    localStorage.removeItem('author');
-    localStorage.removeItem('summary');
-    localStorage.removeItem('bookTitle');
-    localStorage.removeItem('image');
+    removeStoredBookDetails();
   }
 
   /**
@@ -144,8 +116,8 @@ class DetailsForm extends React.Component {
     event.preventDefault();
     this.setState({
       display: true,
+      displayPreloader: 'block'
     });
-    display = 'block';
     this.props.borrowBooks(localStorage.jwtToken, localStorage.bookId);
     $('#modal1').modal('close');
   }
@@ -231,13 +203,13 @@ class DetailsForm extends React.Component {
             <div className="placeholders  ">
               <div className=" col s12 col m2 col l2 placeholder" id="photo">
                 <div id="img_body">
-                  <img src={image} alt="avatar" style={{ width: '150px', height: '250px' }} />
+                  <img src={this.state.filteredData.image} alt="avatar" style={{ width: '150px', height: '250px' }} />
                 </div>
               </div>
             </div>
             <div className=" col s10 col m7 col l7">
 
-              <h5 className="page-header">{bookTitle}</h5>
+              <h5 className="page-header">{this.state.filteredData.bookTitle}</h5>
               <hr />
               <h5>Summary</h5>
               {displayError}
@@ -245,7 +217,7 @@ class DetailsForm extends React.Component {
               <div id="modal1" className="modal">
                 <div className="modal-content">
                   <h5>Do you want to add this book to your archives?</h5>
-                  <p>{bookTitle}</p>
+                  <p>{this.state.filteredData.bookTitle}</p>
                 </div>
                 <div className="modal-footer">
                   <a
@@ -260,11 +232,11 @@ class DetailsForm extends React.Component {
                   >YES</a>
                 </div>
               </div>
-              <span id="sum" className="text-muted">{summary}</span>
+              <span id="sum" className="text-muted">{this.state.filteredData.summary}</span>
               <div className="">
-                <p className="bookinfo">Category: {category} </p>
-                <p className="bookinfo">ISBN: {isbn} </p>
-                <p className="bookinfo">Number in Stock: {stocknumber} </p>
+                <p className="bookinfo">Category: {this.state.filteredData.category} </p>
+                <p className="bookinfo">ISBN: {this.state.filteredData.isbn} </p>
+                <p className="bookinfo">Number in Stock: {this.state.filteredData.stockNumber} </p>
               </div>
               <div className="form-inline">
                 <button
@@ -283,7 +255,7 @@ class DetailsForm extends React.Component {
 
           </div>
         </div>
-        <div style={{ display: display.toString() }} id="activity-loader-id" className="activity">
+        <div style={{ display: this.state.displayPreloader.toString() }} id="activity-loader-id" className="activity">
           <ActivityLoader />
         </div>
       </div>
@@ -298,7 +270,5 @@ DetailsForm.propTypes = {
   bookData: PropTypes.arrayOf(PropTypes.any).isRequired,
   borrowBookItem: PropTypes.arrayOf(PropTypes.any).isRequired,
 };
-
-
 export default DetailsForm;
 
