@@ -2,7 +2,6 @@ import React from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 
-
 import BooksPhoto from './BooksPhoto';
 import SideBar from '../SideBar';
 import BooksFilter from './BooksFilter';
@@ -20,14 +19,29 @@ import showbooksByCategory from '../../actions/showbooksByCategory';
  */
 export class BooksPage extends React.Component {
   /**
+   * Creates an instance of BooksFilter.
+   * @param {object} props
+   * @memberof BooksPage
+   */
+  constructor(props) {
+    super(props);
+    this.state = {
+      filterBy: '',
+      searchText: '',
+      error: '',
+      booksCollectionDisplay: false,
+      bookIsFound: false,
+      sortedData: ''
+    };
+    this.bookFilterHandleChange = this.bookFilterHandleChange.bind(this);
+    this.bookFilterHandleSelected = this.bookFilterHandleSelected.bind(this);
+    this.bookCollectionHandleClick = this.bookCollectionHandleClick.bind(this);
+  }
+  /**
    * @memberof BooksPage
    * @returns {void}
    */
   componentWillMount() {
-    if (!this.props.isFetched) {
-      this.props.getbooks(false);
-    }
-
     if (localStorage.id) {
       localStorage.removeItem('bookId');
       localStorage.removeItem('id');
@@ -39,6 +53,64 @@ export class BooksPage extends React.Component {
       localStorage.removeItem('bookTitle');
       localStorage.removeItem('image');
     }
+  }
+  /**
+ * @memberof BooksPage
+ * @returns {void}
+ */
+  componentDidMount() {
+    if (!this.props.isFetched) {
+      this.props.getbooks(false);
+    }
+  }
+  /**
+  * @function componentWillReceiveProps
+  * @param {object} nextProps
+  * @memberof BooksCategory
+  * @returns {void}
+  */
+  componentWillReceiveProps(nextProps) {
+    if (nextProps.categoryData[0]) {
+      this.setState({
+        sortedData: nextProps.categoryData[0]
+      });
+    }
+  }
+  /**
+  * @description
+  * @param {object} event
+  * @memberof BooksPage
+  * @returns {void}
+  */
+  bookFilterHandleChange(event) {
+    this.setState({ [event.target.name]: event.target.value });
+    if (this.state.filterBy !== '') {
+      this.setState({ error: '', bookIsFound: true });
+      if (this.state.filterBy && this.state.searchText && this.props.bookData) {
+        this.props.searchbooks(this.state.filterBy, this.state.searchText, this.props.bookData);
+        this.setState({ booksCollectionDisplay: true });
+      }
+    }
+  }
+  /**
+   * @param {object} event
+   * @memberof BooksPage
+   * @returns {void}
+   */
+  bookFilterHandleSelected(event) {
+    this.setState({
+      bookIsFound: false,
+      [event.target.name]: event.target.value
+    });
+  }
+  /**
+ * @param {object} event
+ * @param {boolean} [set=true]
+ * @memberof BooksPage
+ * @returns {void}
+ */
+  bookCollectionHandleClick(event) {
+    this.props.checkBookDetails(event.target.name, true);
   }
   /**
    * @returns {views} with siderbar, bookfilter and bookcategory component
@@ -56,25 +128,29 @@ export class BooksPage extends React.Component {
           </div>
           <div id="book_body" className="">
             <div className="row">
-              {this.props.isFetched &&
-                <BooksFilter
-                  checkBookDetails={this.props.checkBookDetails}
-                  searchbooks={this.props.searchbooks}
-                  filteredData={this.props.filteredData}
-                  bookData={this.props.bookData}
-                />}
-              {this.props.isFetched &&
-                <BooksCategory
-                  checkBookDetails={this.props.checkBookDetails}
-                  categoryData={this.props.categoryData}
-                />}
-              {this.props.isFetched &&
-                <BooksPhoto
-                  checkBookDetails={this.props.checkBookDetails}
-                  bookData={this.props.bookData}
-                />}
-              {this.props.isFetched &&
-                <BooksCollection checkBookDetails={this.props.checkBookDetails} heading={'Available books'} bookData={this.props.bookData} />}
+              <BooksFilter
+                state={this.state}
+                handleSelected={this.bookFilterHandleSelected}
+                handleChange={this.bookFilterHandleChange}
+                checkBookDetails={this.bookCollectionHandleClick}
+                searchbooks={this.props.searchbooks}
+                filteredData={this.props.filteredData}
+                bookData={this.props.bookData}
+              />
+              <BooksCategory
+                state={this.state}
+                checkBookDetails={this.bookCollectionHandleClick}
+                categoryData={this.props.categoryData}
+              />
+              <BooksPhoto
+                checkBookDetails={this.props.checkBookDetails}
+                bookData={this.props.bookData}
+              />
+              <BooksCollection
+                heading={'Available books'}
+                bookData={this.props.bookData}
+                checkBookDetails={this.bookCollectionHandleClick}
+              />
             </div>
           </div>
         </div>
@@ -82,7 +158,6 @@ export class BooksPage extends React.Component {
     );
   }
 }
-
 BooksPage.propTypes = {
   categoryData: PropTypes.arrayOf(PropTypes.any).isRequired,
   checkBookDetails: PropTypes.func.isRequired,
@@ -92,9 +167,7 @@ BooksPage.propTypes = {
   isFetched: PropTypes.bool.isRequired,
   searchbooks: PropTypes.func.isRequired,
   showbooksByCategory: PropTypes.func.isRequired
-
 };
-
 /**
  * @param {arrrayOfobject} state
  * @returns {objects} of reducer state
@@ -105,10 +178,8 @@ function mapStateToProps(state) {
     bookData: state.books[0].data,
     categoryData: state.category,
     filteredData: state.getFilteredBooks[0].filteredData,
-
   };
 }
-
 export default connect(mapStateToProps, {
   getbooks,
   searchbooks,
