@@ -8,13 +8,14 @@ import PropTypes from 'prop-types';
 export function BorrowedbooksTableRow(props) {
   const { book } = props;
   const { bookItem } = props;
+  const { image, bookTitle, author, category, isbn } = bookItem[0];
   return (
     <tr >
-      <td><img src={bookItem[0].image} style={{ width: '30px', height: '30px' }} alt="name" /></td>
-      <td>{bookItem[0].bookTitle}</td>
-      <td>{bookItem[0].author}</td>
-      <td>{bookItem[0].category}</td>
-      <td>{bookItem[0].isbn}</td>
+      {image && <td><img src={image} style={{ width: '30px', height: '30px' }} alt="name" /></td>}
+      <td>{bookTitle}</td>
+      <td>{author}</td>
+      <td>{category}</td>
+      <td>{isbn}</td>
       <td>
         <button name={`${book.id},${book.bookId}`} onClick={props.onHandleRead} className="material-icons green-text">import_contacts</button>
         <button name={`${book.id},${book.bookId}`} onClick={props.onHandleClick} className="material-icons red-text">assignment_return</button>
