@@ -3,10 +3,9 @@ import { connect } from 'react-redux';
 import isEmpty from 'lodash/isEmpty';
 import PropTypes from 'prop-types';
 import $ from 'jquery';
-
 import getUserdetails from '../../../actions/getUserDetails';
 import DashboardSidebar from '../DashboardSidebar';
-import ChangePasswordForm from '../ChangePasswordForm';
+import { ChangePasswordForm } from '../ChangePasswordForm';
 import updateUser from '../../../actions/updateuserDetails';
 import { uploadAvatar } from '../../../actions/uploadUserAvatar';
 import changePassword from '../../../actions/changePassword';
@@ -19,6 +18,29 @@ import logout from '../../../actions/logoutAction';
  */
 export class ChangePasswordPage extends React.Component {
   /**
+  * Creates an instance of ChangePasswordForm.
+  * @param {object} props
+  * @memberof ChangePasswordPage
+  */
+  constructor(props) {
+    super(props);
+    this.state = {
+      oldPassword: '',
+      newPassword: '',
+      confirmPassword: '',
+      error: '',
+      display: false,
+      errors: '',
+      message: '',
+      displayPreloader: 'none'
+    };
+    this.handleClick = this.handleClick.bind(this);
+    this.handleInputChange = this.handleInputChange.bind(this);
+    this.handleSave = this.handleSave.bind(this);
+    this.handleClose = this.handleClose.bind(this);
+    this.handleExit = this.handleExit.bind(this);
+  }
+  /**
    * @function compenentWillReceiveProps
    * @memberof ChangePasswordPage
    * @returns {void}
@@ -30,6 +52,110 @@ export class ChangePasswordPage extends React.Component {
     $(document).ready(() => {
       $('.modal').modal();
     });
+  }
+  /**
+  * @param {object} nextProps
+  * @memberof ChangePasswordPage
+  * @returns {void}
+  */
+  componentWillReceiveProps(nextProps) {
+    this.setState({
+      displayPreloader: 'none'
+    });
+    const sortedData = nextProps.passwordChange[0];
+    if (this.state.display) {
+      if (!isEmpty(sortedData.error) && this.state.display) {
+        $('#modalError').modal('open');
+        this.setState({
+          display: false,
+          errors: sortedData.error,
+        });
+      } else if (!isEmpty(sortedData.data) && this.state.display) {
+        this.setState({
+          display: false,
+          oldPassword: '',
+          newPassword: '',
+          confirmPassword: '',
+          message: sortedData.data,
+        });
+        $('#modalSuccess').modal('open');
+      }
+    }
+  }
+  /**
+   * @param {object} event
+   * @memberof ChangePasswordForm
+   * @returns {void}
+   */
+  handleClose(event) {
+    event.preventDefault();
+    $('#modalOpen').modal('close');
+    this.setState({
+      displayErrorMessage: false,
+    });
+  }
+  /**
+   *@returns {void}
+   * @param {object} event
+   * @memberof ChangePasswordForm
+   */
+  handleExit(event) {
+    event.preventDefault();
+
+    $('#modalError').modal('close');
+    $('#modalSuccess').modal('close');
+    this.setState({
+      displayErrorMessage: false,
+    });
+  }
+  /**
+   *@returns {void}
+   * @param {object} event
+   * @memberof ChangePasswordForm
+   */
+  handleClick(event) {
+    event.preventDefault();
+    this.setState({
+      displayPreloader: 'block',
+      display: true,
+    });
+    $('#modalOpen').modal('close');
+    this.props.changePassword(this.state, localStorage.jwtToken);
+  }
+  /**
+   * @param {object} event
+   * @memberof ChangePasswordForm
+   *  @returns {void}
+   */
+  handleInputChange(event) {
+    this.setState({ [event.target.name]: event.target.value });
+  }
+  /**
+   * @param {object} event
+   * @memberof ChangePasswordForm
+   * @returns {void}
+   */
+  handleSave(event) {
+    event.preventDefault();
+    if (!this.state.newPassword || !this.state.oldPassword || !this.state.confirmPassword) {
+      this.setState({
+        error: 'Please enter the required fields',
+      });
+    } else if (this.state.newPassword !== this.state.confirmPassword) {
+      this.setState({
+        error: 'new password and confirm password does not match',
+      });
+    } else if (this.state.newPassword.length < 6) {
+      this.setState({
+        error: 'password length must be more than 5',
+      });
+    } else {
+      this.setState({
+        error: '',
+      });
+
+      $('#modalOpen').modal('open');
+    }
   }
   /**
    * @function render
@@ -51,8 +177,14 @@ export class ChangePasswordPage extends React.Component {
               uploadAvatar={this.props.uploadAvatar}
             />
             <ChangePasswordForm
-              passwordChange={this.props.passwordChange}
-              changePassword={this.props.changePassword}
+              handleClick={this.handleClick}
+              handleClose={this.handleClose}
+              handleExit={this.handleExit}
+              handleInputChange={this.handleInputChange}
+              handleSave={this.handleSave}
+              state={this.state}
+            // passwordChange={this.props.passwordChange}
+            // changePassword={this.props.changePassword}
             />
           </div>
         </div>
@@ -86,7 +218,6 @@ function mapStateToProps(state) {
     error: state.updateUser[0].error.toString(),
     message: state.updateUser[0].data.toString(),
     passwordChange: state.passwordChange,
-
   };
 }
 
