@@ -3,20 +3,16 @@ import isEmpty from 'lodash/isEmpty';
 import PropTypes from 'prop-types';
 import $ from 'jquery';
 import { Link } from 'react-router';
-
 import DoubleActionModal from '../modal/DoubleActionModal';
 import SingleActionModal from '../modal/SingleActionModal';
 import ActivityLoader from '../preloader/ActivityLoader';
 
-
-let imgName = '';
 let displayPreloader = 'none';
-
 /**
  * @class DashboardSidebar
  * @extends {React.Component}
  */
-export class DashboardSidebar extends React.Component {
+class DashboardSidebar extends React.Component {
   /**
    * Creates an instance of DashboardSidebar.
    * @param {object} props
@@ -59,10 +55,6 @@ export class DashboardSidebar extends React.Component {
    */
   componentWillReceiveProps(nextProps) {
     displayPreloader = 'none';
-    if (!isEmpty(nextProps.userData)) {
-      imgName = nextProps.userData[0];
-    }
-
     if (this.state.show) {
       if (!isEmpty(this.props.error)) {
         $('#modaE').modal('open');
@@ -82,12 +74,12 @@ export class DashboardSidebar extends React.Component {
 
     if (this.state.imageloaded) {
       if (!isEmpty(nextProps.imageUrl)) {
+        this.props.userData[0].profileImage = nextProps.imageUrl;
         this.setState({
           imageloaded: false,
           profileImage: nextProps.imageUrl,
           show: true,
         });
-
         this.props.updateUser({ profileImage: nextProps.imageUrl }, localStorage.jwtToken);
       }
     }
@@ -199,9 +191,8 @@ export class DashboardSidebar extends React.Component {
     this.props.logout();
   }
   /**
-   * @returns
    * @memberof DashboardSidebar
-   *  @returns {views} containing side navigation link
+   * @returns {views} containing side navigation link
    */
   render() {
     return (
@@ -214,7 +205,16 @@ export class DashboardSidebar extends React.Component {
                   <div className="file-field input-field">
                     <div id="filebtn" style={{ marginLeft: '5px', color: 'white' }} className="">
                       <span />
-                      <img id="userimg" src={this.state.imagePreviewUrl ? this.state.imagePreviewUrl : !isEmpty(imgName.profileImage) ? imgName.profileImage : 'http://res.cloudinary.com/tosmak/image/upload/v1507297483/userimg_cxeszl.png'} width="120" height="120" alt="images" />
+                      {!_.isEmpty(this.props.userData[0])
+                        && <img
+                          id="userimg"
+                          src={this.state.imagePreviewUrl ?
+                            this.state.imagePreviewUrl :
+                            !isEmpty(this.props.userData[0].profileImage) ? this.props.userData[0].profileImage : 'http://res.cloudinary.com/tosmak/image/upload/v1507297483/userimg_cxeszl.png'}
+                          width="120"
+                          height="120"
+                          alt="images"
+                        />}
                       <input
                         disabled={this.state.disabled}
                         className="fileInput"
@@ -297,7 +297,6 @@ DashboardSidebar.propTypes = {
   message: PropTypes.string.isRequired,
   updateUser: PropTypes.func.isRequired,
   uploadAvatar: PropTypes.func.isRequired
-
 };
 
 export default DashboardSidebar;
