@@ -6,23 +6,21 @@ import PropTypes from 'prop-types';
  * @param {object} props
  * @returns {view} modal
  */
-function SingleActionModal(props) {
-  return (
-    <div id={props.id} className="modal">
-      <div className="modal-content">
-        <h5>{props.heading}</h5>
-        <p>{props.message}</p>
-      </div>
-      <div className="modal-footer">
-        <a
-          href=""
-          onClick={props.onHandleExit}
-          className="modal-action modal-close waves-effect waves-brown btn-flat"
-        >Close</a>
-      </div>
+const SingleActionModal = props => (
+  <div id={props.id} className="modal">
+    <div className="modal-content">
+      <h5>{props.heading}</h5>
+      <p>{props.message}</p>
     </div>
-  );
-}
+    <div className="modal-footer">
+      <a
+        href=""
+        onClick={props.onHandleExit}
+        className="modal-action modal-close waves-effect waves-brown btn-flat"
+      >Close</a>
+    </div>
+  </div>
+);
 SingleActionModal.propTypes = {
   heading: PropTypes.string.isRequired,
   id: PropTypes.string.isRequired,
