@@ -1,9 +1,8 @@
 import users from './users';
 import books from './books';
-import auth from './auth';
+
 
 export {
   users,
   books,
-  auth,
 };
